fix locked boost icon sometimes having the wrong tint (#5152)
Drawables are shared so when applying tint to one the color can change everywhere it is used. This happens here, where the ic_lock_24dp is boost icon & visibility icon in the metadata. The boost icon has the wrong color, it should be lighter to indicate it is not clickable. Mutating the drawable before applying the tint makes sure the color is only change for one and not everywhere. (Note: Usually it is prefered to set the tint on the ImageView that shows the icon, but that is not possible here since the visibility icon is shown in a TextView)

Reviewed-on: https://codeberg.org/tusky/Tusky/pulls/5152
Reviewed-by: Tak <tak@noreply.codeberg.org>
Co-authored-by: Conny Duck <git@connyduck.at>
Co-committed-by: Conny Duck <git@connyduck.at>