use `ActivityResultContracts.PickVisualMedia' for adding media (#5232)
It refactors the original PickMediaFiles to `ActivityResultContracts.PickVisualMedia`
from the Android API. This takes advantage of the Media Picker added in Android 14
enabling easy media selection and "intent-free" media previewing in file manager.
It also allows selecting multiple media in a single click which reduces users'
efforts to interact.

However, it only provides **Visual** media selection, not audio files. Therefore,
`PickMediaFiles` was refactored to `PickAudioFiles` to provide audio selection.

Also, I added the icon for adding audio button and I'm not satisfied with it... Maybe it's a good idea to redo 309f2ab4b2 if merged.

Reviewed-on: https://codeberg.org/tusky/Tusky/pulls/5232
Reviewed-by: connyduck <connyduck@noreply.codeberg.org>
Co-authored-by: William Goodspeed <goodspeed@anche.no>
Co-committed-by: William Goodspeed <goodspeed@anche.no>