Remove id from NotificationSubscribeResult (#5236)
This PR removes the `id` property from the `NotificationSubscribeResult` API model. It turns out that Mastodon and GoToSocial have different types for this field, the former being an `Int` and the latter being a `String`. Since I couldn't find any usage of this field besides API body deserializing, I figured it should be safe to remove.

This is my first code contribution to Tusky, and I haven't worked with Kotlin before, so correct me if I'm wrong on this judgement :)

Closes #5223

Reviewed-on: https://codeberg.org/tusky/Tusky/pulls/5236
Reviewed-by: Tak <tak@noreply.codeberg.org>
Co-authored-by: Nikita Karamov <me@kytta.dev>
Co-committed-by: Nikita Karamov <me@kytta.dev>