New scheduled posts view (#5229)
This is a complete (re-)implementation of the scheduled posts view. It adds all missing features, like previewing media, showing the scheduled date, tablet mode, etc.
Workarounds to GoToSocial quirks and [this bug](https://github.com/mastodon/mastodon/issues/24645) are included for maximum combatibility.
`MediaAttachments` is a reimplementation of the media previews in the timeline with Compose. It has some additional features, like an indication for more than 4 images and we will be able to reuse it for the timelines when we migrate those to Compose as well.

Known limitations:
- there are no transitions when opening the media viewer
- the media viewer will re-download media because Coil and Glide use different caches

Both of those will go away though once the media viewer is Compose as well

- I removed the "Refresh" action from the menu

Closes https://codeberg.org/tusky/Tusky/issues/5179

Reviewed-on: https://codeberg.org/tusky/Tusky/pulls/5229
Reviewed-by: Tak <tak@noreply.codeberg.org>
Co-authored-by: Conny Duck <git@connyduck.at>
Co-committed-by: Conny Duck <git@connyduck.at>