Fix #2980: Add compose button to Tag activity lists (#5275)
I've added a compose button to `StatusListActivity` which is visible if it of `TAG` kind and triggers `ComposeActivity` with the input field prefilled with the hashtag.

I've copied a fair share of code from `MainActivity`, so maybe it would be reasonable to create a separate subclass for `ComposeButton` to contain the common parts? I'm also a bit confused by related SDK version comparison in `MainActivity`, and I'm not sure how it is relevant to window insets so I left it out as I see no difference.

Reviewed-on: https://codeberg.org/tusky/Tusky/pulls/5275
Reviewed-by: connyduck <connyduck@noreply.codeberg.org>
Co-authored-by: Artha <artha@samerion.com>
Co-committed-by: Artha <artha@samerion.com>